BTA-3411 Add URL path to Document model (#180)
Adds a URL value to the Document model, containing a path to the saved
location of the document. This change is made to allow us to access full
PDFs from TradeBFX, rather than image thumbnails as are returned from
the `upload` value.

Linked to: https://azafinance.atlassian.net/browse/BTA-3411